fixed checkstyle

+++ src/main/java/com/example/onlinebookstore/service/impl/ShoppingCartServiceImpl.java
@@ -38,8 +38,9 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
 
     @Override
     public void delete(Long id) {
-        cartItemRepository.findById(id).orElseThrow(() ->
-                new EntityNotFoundException("Can't find Cart Item by id: " + id));
+        if (!cartItemRepository.existsById(id)) {
+            throw new EntityNotFoundException("Can't find Cart Item by id: " + id);
+        }
         cartItemRepository.deleteById(id);
     }
